requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.8
pyautogui>=0.9.50
pyyaml>=5.1
numpy>=1.19.0
pytest>=6.0.0
pyinstaller>=5.0
Pillow>=9.0
